chispart-cloud/requirements.txt
Flask
pypdf
python-dotenv
Flask-SocketIO
boto3
redis
celery
PyYAML
docker
